Move putchar and fd_write only into wasi, not generic wasm
