+++ src/runtime/runtime_wasm_js.go
@@ -60,3 +60,6 @@ func sleepTicks(d timeUnit)
 
 //export runtime.ticks
 func ticks() timeUnit
+
+// placeholder
+func putchar(c byte) {}
